Move all data types to core
