+++ src/contract/factory.rs
@@ -1,7 +1,7 @@
 use std::{path::Path, sync::Arc};
 
 use crate::{
-    crypto::ZilAddress,
+    core::ZilAddress,
     middlewares::Middleware,
     transaction::{TransactionBuilder, TransactionParams},
     util::parse_zil,
