+++ src/core/mod.rs
@@ -1,11 +1,18 @@
-//! Core data types.
+//! Common data types and functionalities.
 
 pub mod bnum;
 pub mod net;
+pub mod private_key;
 pub mod proto;
+pub mod public_key;
 pub mod tx_hash;
 pub mod types;
+pub mod zil_address;
 
 pub use bnum::*;
 pub use tx_hash::*;
 pub use types::*;
+
+pub use private_key::PrivateKey;
+pub use public_key::PublicKey;
+pub use zil_address::ZilAddress;
