+++ src/crypto/mod.rs
@@ -1,9 +1,3 @@
-pub mod private_key;
-pub mod public_key;
 pub mod schnorr;
-pub mod zil_address;
 
 pub use k256::ecdsa::Signature;
-pub use private_key::PrivateKey;
-pub use public_key::PublicKey;
-pub use zil_address::ZilAddress;
