+++ tests/call_transition.rs
@@ -1,8 +1,11 @@
 mod common;
 
-use crate::common::TestContext;
+use crate::core::TestContext;
 use test_context::test_context;
-use zilliqa_rs::{contract, core::BNum, crypto::ZilAddress};
+use zilliqa_rs::{
+    contract,
+    core::{BNum, ZilAddress},
+};
 
 #[test_context(TestContext)]
 #[tokio::test]
